requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.4.0
cohere>=4.0.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0